Remove geocat-ncomp functionality
Functions from geocat.ncomp are imported into the geocat.comp namespace
by using "from geocat.ncomp import *" inside a try/except. If
geocat-ncomp is not installed, then these functions will silently fail
to import.
