+++ geocat/comp/polynomial.py
