--- test/cython_tests/test_cython.pyx
@@ -1,1 +1,0 @@
-include "test_ncomp.pyx"
