--- test/test_import.py
@@ -1,5 +1,0 @@
-import unittest as ut
-
-class Test_Import(ut.TestCase):
-    def test_import(self):
-        import geocat.comp
